index.js 리팩토링 * Handler 함수 외부로 추출
